+++ GhostRunner.DAL.Interface/ITaskScriptDataAccess.cs
@@ -8,7 +8,7 @@ namespace GhostRunner.DAL.Interface
 {
     public interface ITaskScriptDataAccess
     {
-        TaskScript Get(int taskScriptId);
+        TaskScript Get(long taskScriptId);
 
         TaskScript Insert(TaskScript taskScript);
     }
